requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.9.0.post1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.0
openai==1.57.1
packaging==24.2
pydantic==2.10.3
pydantic_core==2.27.1
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
